fix mongodb migrations

+++ internal/location/model/locaiton.go
@@ -4,6 +4,6 @@ type Location struct {
 	Id       int    `json:"id"`
 	DriverId string `json:"driver_id"`
 
-	Lat float32 `json:"lat"`
-	Lng float32 `json:"lng"`
+	Lat float64 `json:"lat"`
+	Lng float64 `json:"lng"`
 }
