+++ internal/location/model/requests/get_drivers_by_location.go
@@ -1,9 +1,9 @@
 package requests
 
 type GetDriversByLocationReqBody struct {
-	Lat    float32 `json:"lat"`
-	Lng    float32 `json:"lng"`
-	Radius float32 `json:"radius"`
+	Lat    float64 `json:"lat"`
+	Lng    float64 `json:"lng"`
+	Radius float64 `json:"radius"`
 }
 
 type GetDriversByLocationResBody struct {
